requirements/base.txt
Django==1.11.8
djangorestframework==3.7.2
django-webpack-loader==0.5.0
gunicorn==19.7.1
mock==2.0.0
psycopg2==2.6.2
requests==2.14.0
structlog==17.2.0
raven==6.1.0
django-cors-headers==2.1.0
Jinja2==2.9.6
MarkupSafe==1.0
six==1.10.0
sqlparse==0.2.3
django-fsm==2.6.0
dj-static==0.0.6
celery==4.1.0
flower==0.9.2